fix: #544 compute across-route basket totals in one pass and drop hidden PDF headers
